+++ Pandora/Pattern/Functor.hs
@@ -14,7 +14,7 @@ import Pandora.Pattern.Functor.Determinable as Exports
 import Pandora.Pattern.Functor.Extractable as Exports
 import Pandora.Pattern.Functor.Pointable as Exports
 import Pandora.Pattern.Functor.Avoidable as Exports
-import Pandora.Pattern.Functor.Applicative as Exports
+import Pandora.Pattern.Functor.Semimonoidal as Exports
 import Pandora.Pattern.Functor.Alternative as Exports
 import Pandora.Pattern.Functor.Invariant as Exports
 import Pandora.Pattern.Functor.Contravariant as Exports