+++ CHANGELOG.md
@@ -533,3 +533,4 @@
 * Use `Semigroupoid` as superclass for `Semimonoidal`
 * Define `*>-` and `forever_` methods of `Applicative_` typeclass
 * Remove `<+*>`, `<**+>`, `<***+>` methods
+* Remove `Applicative` typeclass